Take replay scoreboard and its friends away from replayitem file
